fix: remove coregistrationjobtemplate and move all to CoregistrationSimularity class
