+++ up42/processing_templates.py
@@ -1,6 +1,8 @@
 import dataclasses
 from typing import List, Optional, Union
 
+import pystac
+
 from up42 import base, processing
 
 
@@ -76,9 +78,20 @@ class DetectionChangeSPOTHyperverge(WorkspaceIdMultiItemTemplate):
 
 
 @dataclasses.dataclass
-class CoregistationSimularity(processing.CoregistrationJobTemplate):
+class CoregistationSimularity(processing.JobTemplate):
+    title: str
+    source_item: pystac.Item
+    reference_item: pystac.Item
     process_id = "coregistration-simularity"
     workspace_id: Union[str, base.WorkspaceId] = dataclasses.field(default=base.WorkspaceId())
+
+    @property
+    def inputs(self) -> dict:
+        return {
+            "title": self.title,
+            "sourceItem": self.source_item.get_self_href(),
+            "referenceItem": self.reference_item.get_self_href(),
+        }
 
 
 @dataclasses.dataclass
